Python/requirements.txt
numpy
pyserial
tk
matplotlib
scipy